renamed noiseUtilities.py SignalToNoise.py in anticipation of creating the SignalToNoise module
