+++ python/lsst/sims/photUtils/SignalToNoise.py
